+++ lhc_web/modules/lhchat/deletechatadmin.php
@@ -1,7 +1,5 @@
 <?php
 
-$chat = erLhcoreClassChat::getSession()->load( 'erLhcoreClassModelChat', $Params['user_parameters']['chat_id']);
-
 $currentUser = erLhcoreClassUser::instance();
 
 if (!isset($_SERVER['HTTP_X_CSRFTOKEN']) || !$currentUser->validateCSFRToken($_SERVER['HTTP_X_CSRFTOKEN'])) {
@@ -9,15 +7,27 @@ if (!isset($_SERVER['HTTP_X_CSRFTOKEN']) || !$currentUser->validateCSFRToken($_S
 	exit;
 }
 
-CSCacheAPC::getMem()->removeFromArray('lhc_open_chats', $chat->id);
+$db = ezcDbInstance::get();
+$db->beginTransaction();
 
-if ($currentUser->hasAccessTo('lhchat','deleteglobalchat') || ($currentUser->hasAccessTo('lhchat','deletechat') && $chat->user_id == $currentUser->getUserID()))
-{
-	erLhcoreClassChatEventDispatcher::getInstance()->dispatch('chat.delete',array('chat' => & $chat,'user' => $currentUser));
-	$chat->removeThis();
-    echo json_encode(array('error' => 'false', 'result' => 'ok' ));
-} else {
-   echo json_encode(array('error' => 'true', 'result' => erTranslationClassLhTranslation::getInstance()->getTranslation('chat/deletechatadmin','You do not have rights to delete a chat') ));
+try {
+    $chat = erLhcoreClassModelChat::fetchAndLock($Params['user_parameters']['chat_id']);
+
+    if ($chat->can_edit_chat && ($currentUser->hasAccessTo('lhchat','deleteglobalchat') || ($currentUser->hasAccessTo('lhchat','deletechat') && $chat->user_id == $currentUser->getUserID())))
+    {
+        erLhcoreClassChatEventDispatcher::getInstance()->dispatch('chat.delete',array('chat' => & $chat, 'user' => $currentUser));
+        $chat->removeThis();
+        echo json_encode(array('error' => false, 'result' => 'ok' ));
+    } else {
+        echo json_encode(array('error' => true, 'result' => erTranslationClassLhTranslation::getInstance()->getTranslation('chat/deletechatadmin','You do not have rights to delete a chat') ));
+    }
+
+    $db->commit();
+
+} catch (Exception $e) {
+    erLhcoreClassLog::write($e->getTraceAsString());
+    echo json_encode(array('error' => true, 'result' => $e->getMessage() ));
+    $db->rollback();
 }
 
 exit;
